Point route config link to fastify version 4.8.x
The route config link still points to document for fastify 2.1.x.

This MR modifies the link to 4.8.x (since there is no link for major version `4.x`). It's slightly better than just pointing to `latest` as this plugin might not be compatible with the future fastify version (e.g. v5) right away.